async_fastapi_service/requirements.txt
elasticsearch7[async]==7.17.8
elasticsearch==8.6.0
fastapi==0.89.1
fastapi-cache2==0.2.0
gunicorn==20.1.0
orjson==3.8.5
pydantic==1.10.4
python-dotenv==0.21.0
redis==4.4.2
uvicorn==0.20.0
uvloop==0.17.0
